requirements.txt
anyio==3.7.1
certifi==2023.5.7
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
python-telegram-bot==20.3
sniffio==1.3.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
